requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.3.0
certifi==2021.5.30
charset-normalizer==2.0.3
colorama==0.4.4
coverage==5.5
dj-database-url==0.5.0
Django==3.1.2
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
iniconfig==1.1.1
model-bakery==1.2.1
packaging==20.9
Pillow==8.1.0
pluggy==0.13.1
psycopg2==2.9.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.2.0
python-dateutil==1.5
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
whitenoise==5.3.0
